requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.36
pymysql==1.1.2
python-dotenv==1.0.0
Werkzeug==3.0.1
pytz==2024.1
feedparser==6.0.11
requests==2.32.3
beautifulsoup4==4.12.3
flask_babel